+++ core/src/main/kotlin/com/gitlab/kordlib/core/cache/data/GuildData.kt
@@ -24,7 +24,7 @@ data class GuildData(
         val explicitContentFilter: ExplicitContentFilter,
         val roles: List<Long>,
         val emojis: List<EmojiData>,
-        val features: List<String>,
+        val features: List<GuildFeature>,
         val mfaLevel: MFALevel,
         val applicationId: Long? = null,
         val widgetEnabled: Boolean? = null,
@@ -41,7 +41,10 @@ data class GuildData(
         val maxMembers: Int? = null,
         val vanityUrlCode: String? = null,
         val description: String? = null,
-        val banner: String? = null
+        val banner: String? = null,
+        val systemChannelFlags: SystemChannelFlags? = null,
+        val rulesChannelId: Long? = null,
+        val discoverySplash: String? = null
 ) {
     companion object {
 
@@ -91,7 +94,10 @@ data class GuildData(
                     maxMembers,
                     vanityUrlCode,
                     description,
-                    banner
+                    banner,
+                    systemChannelFlags,
+                    rulesChannelId?.toLong(),
+                    discoverySplash
             )
         }
     }