+++ CHANGELOG.md
@@ -19,12 +19,21 @@ This will be done by default for Kord clients now.
 allows you to more easily share configuration between gateways.
 * Emojis now have nullable names, this only appears when interacting with guild emojis that have been deleted.
 * Rest, Gateway and Common entities have gained a `Discord` prefix to reduce name collisions with Core.
-* `ParallalRequestHandler` has been upgraded to stable
+* `ParallalRequestHandler` has been upgraded to stable.
+* `features` in `Guilds` are now represented as enum values instead of Strings. 
 
 ## Additions
 
 * Added `LiveEntity` and its implementations to Core. These are self-updating entities that contain a filtered
 event stream, only emitting related events.
+* Added `targetUser` and `targetUserType` to invite creation. #47
+* Added a `rules channel`, `SystemChannelFlags` and `discoverySplash` to `Guild`. #48
+* Added `premium since` to `Member`. #45
+* Added `inviter` to `Invite`. #44
+
+## Fixes
+
+* Ending the process while enabling the shutdownHook and without logging in causes UninitializedPropertyAccessException #50
 
 ## Dependencies